+++ packages/@eventual/aws-runtime/src/handlers/entity-stream-worker.ts
@@ -11,6 +11,7 @@ import {
   createEntityStreamWorker,
   getLazy,
   normalizeCompositeKey,
+  removeGeneratedKeyAttributes,
 } from "@eventual/core-runtime";
 import type { EntityStreamOperation } from "@eventual/core/internal";
 import type { DynamoDBStreamHandler } from "aws-lambda";
@@ -84,18 +85,13 @@ export default (async (event) => {
     const normalizedKey = normalizeCompositeKey(entity, bestValue);
     const keyMap = convertNormalizedEntityKeyToMap(normalizedKey);
 
-    if (newValue) {
-      delete newValue[EntityEntityRecord.VERSION_FIELD];
-      delete newValue[normalizedKey.partition.keyAttribute];
-      if (normalizedKey.sort) {
-        delete newValue[normalizedKey.sort.keyAttribute];
-      }
-    }
-    if (oldValue) {
-      delete oldValue[EntityEntityRecord.VERSION_FIELD];
-      delete oldValue[normalizedKey.partition.keyAttribute];
-      if (normalizedKey.sort) {
-        delete oldValue[normalizedKey.sort.keyAttribute];
+    removeSyntheticKey(newValue);
+    removeSyntheticKey(oldValue);
+
+    function removeSyntheticKey(value: Record<string, any> | undefined) {
+      if (value) {
+        delete value[EntityEntityRecord.VERSION_FIELD];
+        removeGeneratedKeyAttributes(entity!, value, false, true);
       }
     }
 
@@ -106,6 +102,7 @@ export default (async (event) => {
         newValue: newValue as any,
         newVersion,
         operation,
+        // @ts-ignore
         oldValue,
         oldVersion,
       };